requirements.txt
mcp
langgraph
chromadb
google-genai
httpx
python-dotenv
pydantic
rich
